requirements.txt
aiodns
dnspython